+++ packages/sdk/python/src/latitude_sdk/types/__init__.py
@@ -36,6 +36,7 @@ if typing.TYPE_CHECKING:
     from .analytics_query import (
         AnalyticsQuery,
         AnalyticsQuery_Behaviors,
+        AnalyticsQuery_Moments,
         AnalyticsQuery_Scores,
         AnalyticsQuery_Sessions,
         AnalyticsQuery_Spans,
@@ -62,6 +63,31 @@ if typing.TYPE_CHECKING:
     from .analytics_query_behaviors_range import AnalyticsQueryBehaviorsRange
     from .analytics_query_behaviors_time_bucket import AnalyticsQueryBehaviorsTimeBucket
     from .analytics_query_behaviors_time_bucket_unit import AnalyticsQueryBehaviorsTimeBucketUnit
+    from .analytics_query_moments import AnalyticsQueryMoments
+    from .analytics_query_moments_breakdown import AnalyticsQueryMomentsBreakdown
+    from .analytics_query_moments_metric import (
+        AnalyticsQueryMomentsMetric,
+        AnalyticsQueryMomentsMetric_Avg,
+        AnalyticsQueryMomentsMetric_Count,
+        AnalyticsQueryMomentsMetric_Max,
+        AnalyticsQueryMomentsMetric_Median,
+        AnalyticsQueryMomentsMetric_Min,
+    )
+    from .analytics_query_moments_metric_avg import AnalyticsQueryMomentsMetricAvg
+    from .analytics_query_moments_metric_avg_field import AnalyticsQueryMomentsMetricAvgField
+    from .analytics_query_moments_metric_count import AnalyticsQueryMomentsMetricCount
+    from .analytics_query_moments_metric_max import AnalyticsQueryMomentsMetricMax
+    from .analytics_query_moments_metric_max_field import AnalyticsQueryMomentsMetricMaxField
+    from .analytics_query_moments_metric_median import AnalyticsQueryMomentsMetricMedian
+    from .analytics_query_moments_metric_median_field import AnalyticsQueryMomentsMetricMedianField
+    from .analytics_query_moments_metric_min import AnalyticsQueryMomentsMetricMin
+    from .analytics_query_moments_metric_min_field import AnalyticsQueryMomentsMetricMinField
+    from .analytics_query_moments_order_by import AnalyticsQueryMomentsOrderBy
+    from .analytics_query_moments_order_by_by import AnalyticsQueryMomentsOrderByBy
+    from .analytics_query_moments_order_by_direction import AnalyticsQueryMomentsOrderByDirection
+    from .analytics_query_moments_range import AnalyticsQueryMomentsRange
+    from .analytics_query_moments_time_bucket import AnalyticsQueryMomentsTimeBucket
+    from .analytics_query_moments_time_bucket_unit import AnalyticsQueryMomentsTimeBucketUnit
     from .analytics_query_scores import AnalyticsQueryScores
     from .analytics_query_scores_breakdown import AnalyticsQueryScoresBreakdown
     from .analytics_query_scores_metric import (
@@ -98,6 +124,7 @@ if typing.TYPE_CHECKING:
         AnalyticsQuerySessionsMetric_Max,
         AnalyticsQuerySessionsMetric_Median,
         AnalyticsQuerySessionsMetric_Min,
+        AnalyticsQuerySessionsMetric_P95,
         AnalyticsQuerySessionsMetric_Sum,
     )
     from .analytics_query_sessions_metric_avg import AnalyticsQuerySessionsMetricAvg
@@ -111,6 +138,8 @@ if typing.TYPE_CHECKING:
     from .analytics_query_sessions_metric_median_field import AnalyticsQuerySessionsMetricMedianField
     from .analytics_query_sessions_metric_min import AnalyticsQuerySessionsMetricMin
     from .analytics_query_sessions_metric_min_field import AnalyticsQuerySessionsMetricMinField
+    from .analytics_query_sessions_metric_p_95 import AnalyticsQuerySessionsMetricP95
+    from .analytics_query_sessions_metric_p_95_field import AnalyticsQuerySessionsMetricP95Field
     from .analytics_query_sessions_metric_sum import AnalyticsQuerySessionsMetricSum
     from .analytics_query_sessions_metric_sum_field import AnalyticsQuerySessionsMetricSumField
     from .analytics_query_sessions_order_by import AnalyticsQuerySessionsOrderBy
@@ -130,6 +159,7 @@ if typing.TYPE_CHECKING:
         AnalyticsQuerySpansMetric_Max,
         AnalyticsQuerySpansMetric_Median,
         AnalyticsQuerySpansMetric_Min,
+        AnalyticsQuerySpansMetric_P95,
         AnalyticsQuerySpansMetric_Sum,
     )
     from .analytics_query_spans_metric_avg import AnalyticsQuerySpansMetricAvg
@@ -143,6 +173,8 @@ if typing.TYPE_CHECKING:
     from .analytics_query_spans_metric_median_field import AnalyticsQuerySpansMetricMedianField
     from .analytics_query_spans_metric_min import AnalyticsQuerySpansMetricMin
     from .analytics_query_spans_metric_min_field import AnalyticsQuerySpansMetricMinField
+    from .analytics_query_spans_metric_p_95 import AnalyticsQuerySpansMetricP95
+    from .analytics_query_spans_metric_p_95_field import AnalyticsQuerySpansMetricP95Field
     from .analytics_query_spans_metric_sum import AnalyticsQuerySpansMetricSum
     from .analytics_query_spans_metric_sum_field import AnalyticsQuerySpansMetricSumField
     from .analytics_query_spans_order_by import AnalyticsQuerySpansOrderBy
@@ -162,6 +194,7 @@ if typing.TYPE_CHECKING:
         AnalyticsQueryTracesMetric_Max,
         AnalyticsQueryTracesMetric_Median,
         AnalyticsQueryTracesMetric_Min,
+        AnalyticsQueryTracesMetric_P95,
         AnalyticsQueryTracesMetric_Sum,
     )
     from .analytics_query_traces_metric_avg import AnalyticsQueryTracesMetricAvg
@@ -175,6 +208,8 @@ if typing.TYPE_CHECKING:
     from .analytics_query_traces_metric_median_field import AnalyticsQueryTracesMetricMedianField
     from .analytics_query_traces_metric_min import AnalyticsQueryTracesMetricMin
     from .analytics_query_traces_metric_min_field import AnalyticsQueryTracesMetricMinField
+    from .analytics_query_traces_metric_p_95 import AnalyticsQueryTracesMetricP95
+    from .analytics_query_traces_metric_p_95_field import AnalyticsQueryTracesMetricP95Field
     from .analytics_query_traces_metric_sum import AnalyticsQueryTracesMetricSum
     from .analytics_query_traces_metric_sum_field import AnalyticsQueryTracesMetricSumField
     from .analytics_query_traces_order_by import AnalyticsQueryTracesOrderBy
@@ -455,6 +490,29 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQueryBehaviorsRange": ".analytics_query_behaviors_range",
     "AnalyticsQueryBehaviorsTimeBucket": ".analytics_query_behaviors_time_bucket",
     "AnalyticsQueryBehaviorsTimeBucketUnit": ".analytics_query_behaviors_time_bucket_unit",
+    "AnalyticsQueryMoments": ".analytics_query_moments",
+    "AnalyticsQueryMomentsBreakdown": ".analytics_query_moments_breakdown",
+    "AnalyticsQueryMomentsMetric": ".analytics_query_moments_metric",
+    "AnalyticsQueryMomentsMetricAvg": ".analytics_query_moments_metric_avg",
+    "AnalyticsQueryMomentsMetricAvgField": ".analytics_query_moments_metric_avg_field",
+    "AnalyticsQueryMomentsMetricCount": ".analytics_query_moments_metric_count",
+    "AnalyticsQueryMomentsMetricMax": ".analytics_query_moments_metric_max",
+    "AnalyticsQueryMomentsMetricMaxField": ".analytics_query_moments_metric_max_field",
+    "AnalyticsQueryMomentsMetricMedian": ".analytics_query_moments_metric_median",
+    "AnalyticsQueryMomentsMetricMedianField": ".analytics_query_moments_metric_median_field",
+    "AnalyticsQueryMomentsMetricMin": ".analytics_query_moments_metric_min",
+    "AnalyticsQueryMomentsMetricMinField": ".analytics_query_moments_metric_min_field",
+    "AnalyticsQueryMomentsMetric_Avg": ".analytics_query_moments_metric",
+    "AnalyticsQueryMomentsMetric_Count": ".analytics_query_moments_metric",
+    "AnalyticsQueryMomentsMetric_Max": ".analytics_query_moments_metric",
+    "AnalyticsQueryMomentsMetric_Median": ".analytics_query_moments_metric",
+    "AnalyticsQueryMomentsMetric_Min": ".analytics_query_moments_metric",
+    "AnalyticsQueryMomentsOrderBy": ".analytics_query_moments_order_by",
+    "AnalyticsQueryMomentsOrderByBy": ".analytics_query_moments_order_by_by",
+    "AnalyticsQueryMomentsOrderByDirection": ".analytics_query_moments_order_by_direction",
+    "AnalyticsQueryMomentsRange": ".analytics_query_moments_range",
+    "AnalyticsQueryMomentsTimeBucket": ".analytics_query_moments_time_bucket",
+    "AnalyticsQueryMomentsTimeBucketUnit": ".analytics_query_moments_time_bucket_unit",
     "AnalyticsQueryScores": ".analytics_query_scores",
     "AnalyticsQueryScoresBreakdown": ".analytics_query_scores_breakdown",
     "AnalyticsQueryScoresMetric": ".analytics_query_scores_metric",
@@ -492,6 +550,8 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQuerySessionsMetricMedianField": ".analytics_query_sessions_metric_median_field",
     "AnalyticsQuerySessionsMetricMin": ".analytics_query_sessions_metric_min",
     "AnalyticsQuerySessionsMetricMinField": ".analytics_query_sessions_metric_min_field",
+    "AnalyticsQuerySessionsMetricP95": ".analytics_query_sessions_metric_p_95",
+    "AnalyticsQuerySessionsMetricP95Field": ".analytics_query_sessions_metric_p_95_field",
     "AnalyticsQuerySessionsMetricSum": ".analytics_query_sessions_metric_sum",
     "AnalyticsQuerySessionsMetricSumField": ".analytics_query_sessions_metric_sum_field",
     "AnalyticsQuerySessionsMetric_Avg": ".analytics_query_sessions_metric",
@@ -501,6 +561,7 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQuerySessionsMetric_Max": ".analytics_query_sessions_metric",
     "AnalyticsQuerySessionsMetric_Median": ".analytics_query_sessions_metric",
     "AnalyticsQuerySessionsMetric_Min": ".analytics_query_sessions_metric",
+    "AnalyticsQuerySessionsMetric_P95": ".analytics_query_sessions_metric",
     "AnalyticsQuerySessionsMetric_Sum": ".analytics_query_sessions_metric",
     "AnalyticsQuerySessionsOrderBy": ".analytics_query_sessions_order_by",
     "AnalyticsQuerySessionsOrderByBy": ".analytics_query_sessions_order_by_by",
@@ -522,6 +583,8 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQuerySpansMetricMedianField": ".analytics_query_spans_metric_median_field",
     "AnalyticsQuerySpansMetricMin": ".analytics_query_spans_metric_min",
     "AnalyticsQuerySpansMetricMinField": ".analytics_query_spans_metric_min_field",
+    "AnalyticsQuerySpansMetricP95": ".analytics_query_spans_metric_p_95",
+    "AnalyticsQuerySpansMetricP95Field": ".analytics_query_spans_metric_p_95_field",
     "AnalyticsQuerySpansMetricSum": ".analytics_query_spans_metric_sum",
     "AnalyticsQuerySpansMetricSumField": ".analytics_query_spans_metric_sum_field",
     "AnalyticsQuerySpansMetric_Avg": ".analytics_query_spans_metric",
@@ -531,6 +594,7 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQuerySpansMetric_Max": ".analytics_query_spans_metric",
     "AnalyticsQuerySpansMetric_Median": ".analytics_query_spans_metric",
     "AnalyticsQuerySpansMetric_Min": ".analytics_query_spans_metric",
+    "AnalyticsQuerySpansMetric_P95": ".analytics_query_spans_metric",
     "AnalyticsQuerySpansMetric_Sum": ".analytics_query_spans_metric",
     "AnalyticsQuerySpansOrderBy": ".analytics_query_spans_order_by",
     "AnalyticsQuerySpansOrderByBy": ".analytics_query_spans_order_by_by",
@@ -552,6 +616,8 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQueryTracesMetricMedianField": ".analytics_query_traces_metric_median_field",
     "AnalyticsQueryTracesMetricMin": ".analytics_query_traces_metric_min",
     "AnalyticsQueryTracesMetricMinField": ".analytics_query_traces_metric_min_field",
+    "AnalyticsQueryTracesMetricP95": ".analytics_query_traces_metric_p_95",
+    "AnalyticsQueryTracesMetricP95Field": ".analytics_query_traces_metric_p_95_field",
     "AnalyticsQueryTracesMetricSum": ".analytics_query_traces_metric_sum",
     "AnalyticsQueryTracesMetricSumField": ".analytics_query_traces_metric_sum_field",
     "AnalyticsQueryTracesMetric_Avg": ".analytics_query_traces_metric",
@@ -561,6 +627,7 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQueryTracesMetric_Max": ".analytics_query_traces_metric",
     "AnalyticsQueryTracesMetric_Median": ".analytics_query_traces_metric",
     "AnalyticsQueryTracesMetric_Min": ".analytics_query_traces_metric",
+    "AnalyticsQueryTracesMetric_P95": ".analytics_query_traces_metric",
     "AnalyticsQueryTracesMetric_Sum": ".analytics_query_traces_metric",
     "AnalyticsQueryTracesOrderBy": ".analytics_query_traces_order_by",
     "AnalyticsQueryTracesOrderByBy": ".analytics_query_traces_order_by_by",
@@ -569,6 +636,7 @@ _dynamic_imports: typing.Dict[str, str] = {
     "AnalyticsQueryTracesTimeBucket": ".analytics_query_traces_time_bucket",
     "AnalyticsQueryTracesTimeBucketUnit": ".analytics_query_traces_time_bucket_unit",
     "AnalyticsQuery_Behaviors": ".analytics_query",
+    "AnalyticsQuery_Moments": ".analytics_query",
     "AnalyticsQuery_Scores": ".analytics_query",
     "AnalyticsQuery_Sessions": ".analytics_query",
     "AnalyticsQuery_Spans": ".analytics_query",
@@ -873,6 +941,29 @@ __all__ = [
     "AnalyticsQueryBehaviorsRange",
     "AnalyticsQueryBehaviorsTimeBucket",
     "AnalyticsQueryBehaviorsTimeBucketUnit",
+    "AnalyticsQueryMoments",
+    "AnalyticsQueryMomentsBreakdown",
+    "AnalyticsQueryMomentsMetric",
+    "AnalyticsQueryMomentsMetricAvg",
+    "AnalyticsQueryMomentsMetricAvgField",
+    "AnalyticsQueryMomentsMetricCount",
+    "AnalyticsQueryMomentsMetricMax",
+    "AnalyticsQueryMomentsMetricMaxField",
+    "AnalyticsQueryMomentsMetricMedian",
+    "AnalyticsQueryMomentsMetricMedianField",
+    "AnalyticsQueryMomentsMetricMin",
+    "AnalyticsQueryMomentsMetricMinField",
+    "AnalyticsQueryMomentsMetric_Avg",
+    "AnalyticsQueryMomentsMetric_Count",
+    "AnalyticsQueryMomentsMetric_Max",
+    "AnalyticsQueryMomentsMetric_Median",
+    "AnalyticsQueryMomentsMetric_Min",
+    "AnalyticsQueryMomentsOrderBy",
+    "AnalyticsQueryMomentsOrderByBy",
+    "AnalyticsQueryMomentsOrderByDirection",
+    "AnalyticsQueryMomentsRange",
+    "AnalyticsQueryMomentsTimeBucket",
+    "AnalyticsQueryMomentsTimeBucketUnit",
     "AnalyticsQueryScores",
     "AnalyticsQueryScoresBreakdown",
     "AnalyticsQueryScoresMetric",
@@ -910,6 +1001,8 @@ __all__ = [
     "AnalyticsQuerySessionsMetricMedianField",
     "AnalyticsQuerySessionsMetricMin",
     "AnalyticsQuerySessionsMetricMinField",
+    "AnalyticsQuerySessionsMetricP95",
+    "AnalyticsQuerySessionsMetricP95Field",
     "AnalyticsQuerySessionsMetricSum",
     "AnalyticsQuerySessionsMetricSumField",
     "AnalyticsQuerySessionsMetric_Avg",
@@ -919,6 +1012,7 @@ __all__ = [
     "AnalyticsQuerySessionsMetric_Max",
     "AnalyticsQuerySessionsMetric_Median",
     "AnalyticsQuerySessionsMetric_Min",
+    "AnalyticsQuerySessionsMetric_P95",
     "AnalyticsQuerySessionsMetric_Sum",
     "AnalyticsQuerySessionsOrderBy",
     "AnalyticsQuerySessionsOrderByBy",
@@ -940,6 +1034,8 @@ __all__ = [
     "AnalyticsQuerySpansMetricMedianField",
     "AnalyticsQuerySpansMetricMin",
     "AnalyticsQuerySpansMetricMinField",
+    "AnalyticsQuerySpansMetricP95",
+    "AnalyticsQuerySpansMetricP95Field",
     "AnalyticsQuerySpansMetricSum",
     "AnalyticsQuerySpansMetricSumField",
     "AnalyticsQuerySpansMetric_Avg",
@@ -949,6 +1045,7 @@ __all__ = [
     "AnalyticsQuerySpansMetric_Max",
     "AnalyticsQuerySpansMetric_Median",
     "AnalyticsQuerySpansMetric_Min",
+    "AnalyticsQuerySpansMetric_P95",
     "AnalyticsQuerySpansMetric_Sum",
     "AnalyticsQuerySpansOrderBy",
     "AnalyticsQuerySpansOrderByBy",
@@ -970,6 +1067,8 @@ __all__ = [
     "AnalyticsQueryTracesMetricMedianField",
     "AnalyticsQueryTracesMetricMin",
     "AnalyticsQueryTracesMetricMinField",
+    "AnalyticsQueryTracesMetricP95",
+    "AnalyticsQueryTracesMetricP95Field",
     "AnalyticsQueryTracesMetricSum",
     "AnalyticsQueryTracesMetricSumField",
     "AnalyticsQueryTracesMetric_Avg",
@@ -979,6 +1078,7 @@ __all__ = [
     "AnalyticsQueryTracesMetric_Max",
     "AnalyticsQueryTracesMetric_Median",
     "AnalyticsQueryTracesMetric_Min",
+    "AnalyticsQueryTracesMetric_P95",
     "AnalyticsQueryTracesMetric_Sum",
     "AnalyticsQueryTracesOrderBy",
     "AnalyticsQueryTracesOrderByBy",
@@ -987,6 +1087,7 @@ __all__ = [
     "AnalyticsQueryTracesTimeBucket",
     "AnalyticsQueryTracesTimeBucketUnit",
     "AnalyticsQuery_Behaviors",
+    "AnalyticsQuery_Moments",
     "AnalyticsQuery_Scores",
     "AnalyticsQuery_Sessions",
     "AnalyticsQuery_Spans",